+++ src/Oro/Bundle/SSOBundle/Security/OAuthToken.php
@@ -3,10 +3,13 @@
 namespace Oro\Bundle\SSOBundle\Security;
 
 use HWI\Bundle\OAuthBundle\Security\Core\Authentication\Token\OAuthToken as HWIOAuthToken;
-use Oro\Bundle\SecurityBundle\Authentication\Token\OrganizationContextTokenInterface;
-use Oro\Bundle\SecurityBundle\Authentication\Token\OrganizationContextTokenSerializerTrait;
+use Oro\Bundle\SecurityBundle\Authentication\Token\OrganizationAwareTokenInterface;
+use Oro\Bundle\SecurityBundle\Authentication\Token\OrganizationAwareTokenTrait;
 
-class OAuthToken extends HWIOAuthToken implements OrganizationContextTokenInterface
+/**
+ * The OAuth authentication token.
+ */
+class OAuthToken extends HWIOAuthToken implements OrganizationAwareTokenInterface
 {
-    use OrganizationContextTokenSerializerTrait;
+    use OrganizationAwareTokenTrait;
 }