+++ src/Oro/Bundle/SecurityBundle/Authentication/Provider/OrganizationRememberMeAuthenticationProvider.php
@@ -2,20 +2,26 @@
 
 namespace Oro\Bundle\SecurityBundle\Authentication\Provider;
 
-use Oro\Bundle\SecurityBundle\Authentication\Guesser\UserOrganizationGuesser;
+use Oro\Bundle\OrganizationBundle\Entity\Organization;
+use Oro\Bundle\SecurityBundle\Authentication\Guesser\OrganizationGuesserInterface;
 use Oro\Bundle\SecurityBundle\Authentication\Token\OrganizationRememberMeTokenFactoryInterface;
+use Oro\Bundle\UserBundle\Entity\AbstractUser;
 use Oro\Bundle\UserBundle\Entity\User;
 use Symfony\Component\Security\Core\Authentication\Provider\RememberMeAuthenticationProvider;
 use Symfony\Component\Security\Core\Authentication\Token\TokenInterface;
 use Symfony\Component\Security\Core\Exception\AuthenticationException;
 use Symfony\Component\Security\Core\Exception\BadCredentialsException;
 
+/**
+ * The authentication provider to retrieve the user and the organization for a OrganizationRememberMeToken.
+ */
 class OrganizationRememberMeAuthenticationProvider extends RememberMeAuthenticationProvider
 {
-    /**
-     * @var OrganizationRememberMeTokenFactoryInterface
-     */
-    protected $tokenFactory;
+    /** @var OrganizationRememberMeTokenFactoryInterface */
+    private $tokenFactory;
+
+    /** @var OrganizationGuesserInterface */
+    private $organizationGuesser;
 
     /**
      * @param OrganizationRememberMeTokenFactoryInterface $tokenFactory
@@ -23,6 +29,14 @@ class OrganizationRememberMeAuthenticationProvider extends RememberMeAuthenticat
     public function setTokenFactory(OrganizationRememberMeTokenFactoryInterface $tokenFactory)
     {
         $this->tokenFactory = $tokenFactory;
+    }
+
+    /**
+     * @param OrganizationGuesserInterface $organizationGuesser
+     */
+    public function setOrganizationGuesser(OrganizationGuesserInterface $organizationGuesser)
+    {
+        $this->organizationGuesser = $organizationGuesser;
     }
 
     /**
@@ -35,31 +49,46 @@ class OrganizationRememberMeAuthenticationProvider extends RememberMeAuthenticat
                 'Token Factory is not set in OrganizationRememberMeAuthenticationProvider.'
             );
         }
+        if (null === $this->organizationGuesser) {
+            throw new AuthenticationException(
+                'Organization Guesser is not set in OrganizationRememberMeAuthenticationProvider.'
+            );
+        }
 
-        $guesser = new UserOrganizationGuesser();
         /**  @var TokenInterface $token */
         $authenticatedToken = parent::authenticate($token);
 
         /** @var User $user */
-        $user         = $authenticatedToken->getUser();
-        $organization = $guesser->guess($user, $token);
+        $user = $authenticatedToken->getUser();
+        $organization = $this->guessOrganization($user, $token);
 
-        if (!$organization) {
-            throw new BadCredentialsException("You don't have active organization assigned.");
-        } elseif (!$user->getOrganizations(true)->contains($organization)) {
-            throw new BadCredentialsException(
-                sprintf("You don't have access to organization '%s'", $organization->getName())
-            );
+        return $this->tokenFactory->create(
+            $user,
+            $authenticatedToken->getProviderKey(),
+            $authenticatedToken->getSecret(),
+            $organization
+        );
+    }
+
+    /**
+     * @param AbstractUser   $user
+     * @param TokenInterface $token
+     *
+     * @return Organization
+     */
+    private function guessOrganization(AbstractUser $user, TokenInterface $token): Organization
+    {
+        $organization = $this->organizationGuesser->guess($user, $token);
+        if (null === $organization) {
+            throw new BadCredentialsException('The user does not have active organization assigned to it.');
+        }
+        if (!$user->isBelongToOrganization($organization, true)) {
+            throw new BadCredentialsException(sprintf(
+                'The user does not have access to organization "%s".',
+                $organization->getName()
+            ));
         }
 
-        $authenticatedToken = $this->tokenFactory
-            ->create(
-                $user,
-                $authenticatedToken->getProviderKey(),
-                $authenticatedToken->getSecret(),
-                $organization
-            );
-
-        return $authenticatedToken;
+        return $organization;
     }
 }